src/requirements.txt
keras
tensorflow
numpy
GitPython